init section table redesign
